Split the parser into multiple files
