+++ src/parsing/parsingFunctions.ts
@@ -1,17 +1,12 @@
 import {
-	complete,
 	delimited,
 	error,
-	followed,
 	map,
-	not,
 	okWithValue,
 	optional,
 	or,
 	Parser,
 	ParseResult,
-	peek,
-	plus,
 	preceded,
 	star,
 	then,
